🪃🐡 ↝ Lightkurve files & stats now being returned to client in #32 #29 #28, generating planet images and prepping to upload them to IPFS/supa for #18 for planet pages in #16
