+++ views/__init__.py
@@ -1,26 +1,20 @@
-from flask import Flask, Blueprint, jsonify
+from flask import Flask, Blueprint, jsonify, request, Response
+from .datastore import supabase, find_all_planets, add_planet_to_DB, planets
 # from .app import app
 # from views.models import Planet
 
 # views.py
-main = Blueprint('main', __name__)
-
-@main.route('/planet', methods=['GET', 'POST'])
-def index():
-    return jsonify(planets)
-
-@main.route('/time')
-def get_current_time():
-    return {'time': time.time()}
+classify = Blueprint('classify', __name__)
 
 # GET request -> return all planets in storage/db in JSON format
-@main.route('/planets')
+@classify.route('/planets')
 def get_planets():
+    planets = find_all_planets()
     return jsonify({
         'planets': planets, # Then present this on react frontend, port 5000 -> 3000
     })
 
-@main.route('/planets/<planet_id>')
+@classify.route('/planets/<planet_id>')
 def find_planet_by_id(planet_id):
     for planet in planets:
         if planet["id"] == int(planet_id):
@@ -28,18 +22,18 @@ def find_planet_by_id(planet_id):
                 "planet": planet,
             })
 
-@main.route('/planets/add', methods=['POST'])
+@classify.route('/planets/add', methods=['POST'])
 def add_planet():
     data = request.get_json()
     try:
         title = data['title']
-        if title:
-            data = add_planet_to_DB(title)
-            return jsonify(data), 201
+        ticId = data['ticId']
+        data = add_planet_to_DB(title, ticId)
+        return jsonify(data), 201
     except:
         return Response('''{"message": "Bad Request"}''', status=400, mimetype='application/json')
 
-@main.route('/generator')
+@classify.route('/generator')
 def generate():
     # Generate the figure **without using pyplot**.
     res = 2048 # see generate blueprint above
